web_display/requirements.txt
flask
grpcio
protobuf
prometheus_client